Adding favicon code and initial homepage boilerplate Empty template files for logged-in and logged-out structure for the homepage, and static favicon images.
Relying on the 'default' filenames for the 32x32 and apple icons.

Only defining the 'everyone else modern' favicon to reduce the
reply size as much as possible.

Also removed the '...' trailer on all 'Registering' lines due to
the clutter it generated on startup with paths including filename
extensions such as the favicons.