Use a central 'shared data' object
This simplifies the code long-term by allowing the endpoints to
set additional data that needs to be shared elsewhere, in effect
a constrained global variablespace that is freshly cloned for
each request processed.

Uses the JSON.parse(JSON.stringify(object)) method, so it ONLY
clones JSON-compatible values, which is the intent.

Note this is NOT used for page-specific data such as API keys used
for OAuth or the like, this system is a cross-route-component info
sharing mechanism.