Reducing the IV to the minimal effective size The GHASH component of GCM is specified to short-circuit it's IV when that is 96 bits or less, and zero-extend it.
However if you send it something longer than 96 bits it will use
that given IV and the key as internal seeds to the GHAS systems,
to generate a full 128-bit internal IV it uses.

As such, to maximize the return on computation time, using a 128-
bit IV will guarantee maximum randomness and minimize any chance
of hash collissions without wating CPU needlessly.