Splitting out 'routes' and hardening cookies
The 'routes/' directory can be directly 'require'ed now.

Adjusted the session cookies to strip fixed plaintext caused by
using JSON for packaging, and adding 32 random bytes at the front
AND end of the string before encoding.

This does not affect the underlying aead component, as we're only
'wrapping' the value we hand to it in the new data.