Splitting util.js up w/ code coverage
Added support for Istanbul-based code coverage tests.
Current the login_complete function has full code coverage,
and as a side-effect the database and keyvalue were reviewed
as well. Though most 'throw error' fragments are marked to be
ignored by istanbul currently as it's nigh-impossible to rig
up such errors to occur on demand.