Adding BattleNet and InstaGram
Add to add a few more components to the OAuth2 to cover all bases.

For many services it's sending the ClientID/ClientSecret THREE
ways at once now on requests, just to cover all bases without
adding a bajillion flags and switches.

Had to update the oauth2 tests to use string-prefix tests on the
mini-HTTPS server they use as a faux-API server to contact.